serializer: serialize recursive structures
Fix bug with bus error during serializing of recursive structures
and always print aliases if node become referenced before or
after serialization.

This patch adds map with serialized nodes and results. The map is
used to mark referenced nodes which appears during serializing.

Closes #3228

@TarantoolBot
__serialize parameter
If __serialize parameter is initialized as a function then this
function should be pure. A pure function is a function which
returns identical values for identical arguments (no variation
with local static variables, non-local variables, mutable
reference arguments or input streams). Otherwise, the behavior
is undefined.